refactor(site): lift nav open state to AppLayout and co-locate MenuItem under SideNavigation
Move open/close state, scroll lock, and resize listener from SideNavigation to
AppLayout so the header stays always visible while the nav panel slides
independently. Relocate MenuItem from Layout/MenuItem/ to
SideNavigation/MenuItem/ and update all import paths and test mocks. Fix
vitest list ESM resolution for next-intl by adding server.deps.inline.
