+++ apps/site/src/components/Layout/AppLayout/index.tsx
@@ -1,31 +1,41 @@
 'use client';
 
-import { FC, PropsWithChildren } from 'react';
+import { FC, PropsWithChildren, useCallback } from 'react';
 
-// import { useBoolean, useEventListener } from 'usehooks-ts';
+import { useBoolean, useScrollLock, useEventListener } from 'usehooks-ts';
 
-// import { useThrottle } from '~hooks';
-// import { BREAKPOINTS_NUMBERS } from '~utils';
+import { useThrottle } from '~hooks';
+import { BREAKPOINTS_NUMBERS } from '~utils';
 
 import { SideNavigation } from '../SideNavigation';
 
 export const AppLayout: FC<PropsWithChildren> = ({ children }) => {
-  // const throttle = useThrottle(() => {
-  //   close();
-  // }, 500);
+  const { value, setTrue, setFalse } = useBoolean(false);
+  const { lock, unlock } = useScrollLock({ autoLock: false });
 
-  // const value = useMemo(() => ({ open, toggle, close }), [open, toggle, close]);
+  const handleToggle = useCallback(() => {
+    if (value) {
+      setFalse();
+      unlock();
+      return;
+    }
 
-  // useEventListener('resize', () => {
-  //   if (window.innerWidth < BREAKPOINTS_NUMBERS.lg) throttle();
-  // });
+    setTrue();
+    lock();
+  }, [value, lock, unlock, setFalse, setTrue]);
 
-  // useBodyClass(classNames({ 'overflow-hidden': open }));
+  const throttle = useThrottle(() => {
+    handleToggle();
+  }, 500);
+
+  useEventListener('resize', () => {
+    if (window.innerWidth < BREAKPOINTS_NUMBERS.lg) throttle();
+  });
 
   return (
     <>
-      <SideNavigation />
-      <div className="flex h-full flex-col ml-0 xl:ml-60 xl:mt-20">
+      <SideNavigation open={value} toggle={handleToggle} />
+      <div className="flex h-full flex-col ml-0 xl:ml-60 mt-header-mobile xl:mt-20">
         <main className="w-full max-w-237.5 2xl:max-w-278.5 mx-auto h-full flex flex-col flex-1">
           {children}
         </main>
