+++ apps/site/src/components/Layout/SideNavigation/index.tsx
@@ -1,45 +1,32 @@
 'use client';
 
-import { FC, useCallback } from 'react';
+import { FC } from 'react';
 
 import { Divider } from '@repo/ui/View';
 import classNames from 'classnames';
 import { useTranslations } from 'next-intl';
-import { useBoolean, useEventListener, useScrollLock } from 'usehooks-ts';
-
-import { useThrottle } from '~hooks';
-import { BREAKPOINTS_NUMBERS } from '~utils';
 
 import { Header } from '../Header';
-import { MenuItem } from '../MenuItem';
 import { LanguageSelector } from './LanguageSelector';
+import { MenuItem } from './MenuItem';
 import { ThemeToggle } from './ThemeToggle';
 
-export const SideNavigation: FC = () => {
+interface SideNavigationProps {
+  open: boolean;
+  toggle: () => void;
+}
+
+export const SideNavigation: FC<SideNavigationProps> = ({ open, toggle }) => {
   const t = useTranslations('SideNavigation');
-  const { lock, unlock } = useScrollLock({ autoLock: false });
-  const { value, setTrue, toggle } = useBoolean(false);
-
-  const throttle = useThrottle(() => {
-    setTrue();
-  }, 500);
-
-  useEventListener('resize', () => {
-    if (window.innerWidth < BREAKPOINTS_NUMBERS.lg) throttle();
-  });
 
   return (
-    <nav
-      id="side-navigation"
-      className={classNames(
-        'w-full xl:w-60 h-screen fixed top-0 z-50 left-0 xl:px-4 bg-surface-sunken flex flex-col border-0 shadow-1 box-content',
-      )}
-    >
-      <Header open={value} toggle={toggle} />
-      <main
+    <section className="fixed top-0 left-0 shadow-1 w-full xl:w-auto z-50">
+      <Header open={open} toggle={toggle} />
+      <nav
+        id="side-navigation"
         className={classNames(
-          'transition-all duration-300 ease-linear h-full xl:!translate-x-0',
-          value ? 'translate-x-0' : '-translate-x-full',
+          'h-sidenav-mobile xl:h-sidenav-desktop left-0 w-full xl:w-60 xl:px-4 bg-surface-sunken flex flex-col border-0 duration-300 ease-linear xl:!translate-x-0 z-9999',
+          open ? 'translate-x-0' : '-translate-x-full',
         )}
       >
         <ul className="flex flex-col gap-y-3 px-6 pt-10 xl:pt-15 xl:px-0">
@@ -91,7 +78,7 @@ export const SideNavigation: FC = () => {
           <ThemeToggle />
           <LanguageSelector />
         </ul>
-      </main>
-    </nav>
+      </nav>
+    </section>
   );
 };
