+++ apps/site/src/components/Layout/index.ts
@@ -1,4 +1,1 @@
 export * from './AppLayout';
-export * from './Header';
-export * from './MenuItem';
-export * from './SideNavigation';
