+++ apps/site/src/features/contact/ContactInfo/index.tsx
@@ -7,7 +7,7 @@ import { Divider } from '@repo/ui/View';
 import { formatCellphone } from '@repo/utils';
 import { useTranslations } from 'next-intl';
 
-import { MenuItem } from '~components/Layout/MenuItem';
+import { MenuItem } from '~/components/Layout/SideNavigation/MenuItem';
 import { Link } from '~i18n/routing';
 
 export const ContactInfo: FC = () => {
