+++ apps/site/tests/components/View/SideNavigation/ThemeToggle.test.tsx
@@ -17,7 +17,7 @@ vi.mock('next-intl', () => ({
   useTranslations: () => (key: string) => key,
 }));
 
-vi.mock('~/components/Layout/MenuItem/index', () => ({
+vi.mock('~/components/Layout/SideNavigation/MenuItem', () => ({
   MenuItem: {
     Item2: {
       Expandable: ({ children }: { children: React.ReactNode }) => (
